# docs/swagger.yaml
basePath: /
definitions:
  handler.request:
    properties:
      arguments:
        additionalProperties: true
        type: object
      tool_name:
        type: string
    type: object
  server.CallToolRequestExample:
    properties:
      name:
        example: tool_name
        type: string
      params: {}
    type: object
  utils.Meta:
    properties:
      execution_time_ms:
        type: integer
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
host: localhost:8081
info:
  contact: {}
  description: Simple HTTP server exposing MCP endpoints
  title: MCP Server API
  version: "1.0"
paths:
  /api/execute:
    post:
      consumes:
      - application/json
      description: Process MCP CallToolRequest and route to correct handler
      parameters:
      - description: MCP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Tool not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Handle MCP tool call
  /health:
    get:
      description: Return server status and version
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
  /mcp:
    post:
      consumes:
      - application/json
      description: Process MCP CallToolRequest and route to correct handler
      parameters:
      - description: MCP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.CallToolRequestExample'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Tool not found
          schema:
            type: string
      summary: Handle MCP tool call
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload new OpenAPI spec file and reload tools
      parameters:
      - description: OpenAPI spec file
        in: formData
        name: spec
        required: true
        type: file
      responses:
        "200":
          description: Upload successful
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Upload OpenAPI spec
securityDefinitions:
  ApiKeyAuth:
    description: Provide your API key here to access the endpoints
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"